+++ profile/index.html
@@ -60,6 +60,14 @@
                     Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse voluptate, laudantium, corrupti consectetur repellat eaque rem autem ipsum
                     nobis placeat odit ut aliquid accusantium quod earum, animi dicta quaerat! Eaque!
                   </p>
+
+                <div>
+                  <h2>My friends</h2>
+                  <p class="fw-bold">Followers</p>
+                  <div class="d-flex flex-lg-column" id="followers"></div>
+                  <p class="fw-bold">Following</p>
+                  <div class="d-flex flex-lg-column" id="following"></div>
+                </div>
                 </div>
               </div>
             </div>
@@ -85,16 +93,7 @@
              </div>
            
             </div>
-            <!--Followers-->
-            <div class="col-lg-3 mb-4">
-              <div class="bg-light rounded-2 p-3">
-                <h2>My friends</h2>
-                <p class="fw-bold">Followers</p>
-                <div class="d-flex flex-lg-column" id="followers"></div>
-                <p class="fw-bold">Following</p>
-                <div class="d-flex flex-lg-column" id="following"></div>
-              </div>
-            </div>
+    
           </div>
         </div>
       </main>
